requirements.txt
amqp==5.0.6
anyjson==0.3.3
arabic-reshaper==2.1.3
asgiref==3.3.4
beautifulsoup4==4.9.3
billiard==3.6.4.0
celery==5.0.0
certifi==2021.5.30
charset-normalizer==2.0.1
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2
django-import-export==2.5.0
django-redis==5.0.0
et-xmlfile==1.1.0
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
idna==3.2
kombu==5.1.0
librabbitmq==2.0.0
MarkupPy==1.14
MouseInfo==0.1.3
numpy==1.20.3
odfpy==1.4.1
openpyxl==3.0.7
Pillow==8.2.0
prompt-toolkit==3.0.19
psycopg2==2.8.6
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
PyPDF2==1.26.0
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.27
python-bidi==0.4.2
python-decouple==3.4
PyTweening==1.0.3
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
reportlab==3.5.67
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
static3==0.7.0
tablib==3.0.0
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.2.0
wikipedia==1.4.0
xhtml2pdf==0.2.5
xlrd==2.0.1
xlwt==1.3.0
